ajustado scollview de listview e separado componentes
