+++ app/src/main/java/com/example/myshoppinglist/screen/HomeScreen.kt
@@ -3,8 +3,6 @@ package com.example.myshoppinglist.screen
 import android.content.Context
 import androidx.compose.animation.ExperimentalAnimationApi
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
 import androidx.compose.material.Text
@@ -12,23 +10,17 @@ import androidx.compose.material.contentColorFor
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.scale
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalLifecycleOwner
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.compose.ui.zIndex
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.MutableLiveData
 import androidx.navigation.NavController
 import com.example.myshoppinglist.R
 import com.example.myshoppinglist.callback.Callback
 import com.example.myshoppinglist.callback.VisibleCallback
-import com.example.myshoppinglist.components.BaseAnimationComponent
-import com.example.myshoppinglist.components.BoxCardCreditCustom
-import com.example.myshoppinglist.components.BoxPurchaseHistoryComponent
-import com.example.myshoppinglist.components.HeaderComponent
+import com.example.myshoppinglist.components.*
 import com.example.myshoppinglist.database.dtos.CreditCardDTODB
 import com.example.myshoppinglist.database.entities.relations.PurchaseAndCategory
 import com.example.myshoppinglist.database.sharedPreference.UserLoggedShared
@@ -40,7 +32,6 @@ import com.example.myshoppinglist.ui.theme.LatoBold
 import com.example.myshoppinglist.ui.theme.text_secondary
 import com.example.myshoppinglist.utils.MountStructureCrediCard
 import com.google.accompanist.pager.ExperimentalPagerApi
-import java.lang.Math.abs
 
 @ExperimentalPagerApi
 @ExperimentalAnimationApi
@@ -96,7 +87,7 @@ fun HomeScreen(navController: NavController?) {
                 }
             })
 
-            Carousel(
+            CarouselComponent(
                 visibleAnimation = visibleAnimation.value,
                 count = creditCardCollection.size,
                 parentModifier = Modifier
@@ -127,7 +118,6 @@ fun HomeScreen(navController: NavController?) {
             Spacer(Modifier.size(24.dp))
 
             BoxPurchaseHistoryComponent(
-                visibleAnimation.value,
                 purchaseCollection,
                 object : VisibleCallback {
                     override fun onChangeVisible(visible: Boolean) {
@@ -139,69 +129,6 @@ fun HomeScreen(navController: NavController?) {
         }
     }
 
-}
-
-@OptIn(ExperimentalAnimationApi::class)
-@Composable
-fun Carousel(
-    count: Int,
-    visibleAnimation: Boolean = true,
-    parentModifier: Modifier = Modifier.fillMaxWidth(),
-    contentHeight: Dp,
-    content: @Composable (modifier: Modifier, index: Int) -> Unit
-) {
-
-    val listState = rememberLazyListState(count / 2)
-
-    LaunchedEffect(key1 = count) {
-        listState.animateScrollToItem(count)
-    }
-
-    BaseAnimationComponent(
-        visibleAnimation = visibleAnimation,
-        contentBase = {
-            BoxWithConstraints(
-                modifier = parentModifier
-            ) {
-                val halfRowWidth = constraints.maxWidth / 2
-
-                LazyColumn(
-                    state = listState,
-                    modifier = Modifier.fillMaxSize(),
-                    contentPadding = PaddingValues(),
-                    verticalArrangement = Arrangement.spacedBy(-contentHeight / 3f),
-                    horizontalAlignment = Alignment.CenterHorizontally
-                ) {
-                    items(
-                        count = count,
-                        itemContent = { globalIndex ->
-
-                            val scale by remember {
-                                derivedStateOf {
-                                    val currentItem =
-                                        listState.layoutInfo.visibleItemsInfo.firstOrNull { it.index == globalIndex }
-                                            ?: return@derivedStateOf 0.85f
-
-                                    (1f - minOf(
-                                        1f,
-                                        abs(currentItem.offset + (currentItem.size / 2) - halfRowWidth).toFloat() / halfRowWidth
-                                    ) * 0.3f)
-                                }
-                            }
-
-                            content(
-                                index = globalIndex,
-                                modifier = Modifier
-                                    .fillMaxWidth()
-                                    .height(contentHeight)
-                                    .scale(scale)
-                                    .zIndex(scale)
-                            )
-                        }
-                    )
-                }
-            }
-        })
 }
 
 class HomeFieldViewModel(context: Context, lifecycleOwner: LifecycleOwner) : BaseFieldViewModel() {
